requirements.txt
streamlit>=1.0.0
google-generativeai>=0.0.1
python-dotenv>=0.20.0
